+++ packages/map-template/src/atoms/showRoadNameLabels.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-const showRoadNameLabelsState = atom({
-    key: 'showRoadNameLabels',
-    default: undefined
-});
-
-export default showRoadNameLabelsState;